prevent modem-manager from grabbing device
This is not a modem, but modem-manager was grabbing the device, and
ignoring my udev rules.  Setting ID_MM_DEVICE_IGNORE to something
apparently will cause modem-manager to ignore the device.

Also updated the stanza to match what the system does.
While this still creates an AsantePorter device, the number following
it not unique to Asante devices anymore, but I might not care.